# services/control-plane/openapi.yaml
openapi: 3.0.3
info:
  title: SRE Agent Control Plane
  version: 1.0.0
  description: Read APIs and chaos trigger for the autonomous SRE demo.
servers:
  - url: http://localhost:3000
paths:
...
  /autonomy:
    get:
      summary: Whether the scheduled/autonomous chaos trigger is currently paused
      responses:
        '200':
          description: Current autonomy state
          content:
            application/json:
              schema:
                type: object
                properties:
                  autonomousChaosPaused:
                    type: boolean
        '500':
          $ref: '#/components/responses/Error'
    post:
      summary: >-
        Pause or resume the scheduled/autonomous chaos trigger. A manual
        Break It call always runs regardless of this setting.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [paused]
              properties:
                paused:
                  type: boolean
      responses:
        '200':
          description: Updated autonomy state
          content:
            application/json:
              schema:
                type: object
                properties:
                  autonomousChaosPaused:
                    type: boolean
        '400':
          $ref: '#/components/responses/Error'
        '500':
          $ref: '#/components/responses/Error'
...
components:
...
  responses:
    Error:
      description: Error payload
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string